+++ src/resolvers/project.rs
@@ -1,6 +1,6 @@
 use crate::db::schema::users::columns;
 use crate::db::schema::{projects, submissions, users};
-use crate::db::orm::{Project, ProjectInsert, Submission, SubmissionInsert, User};
+use crate::db::models::{Project, ProjectInsert, Submission, SubmissionInsert, User};
 use crate::utils::*;
 use diesel::dsl::any;
 use diesel::prelude::*;
