+++ src/types/mod.rs
@@ -5,5 +5,5 @@ mod tokenized;
 pub use token::*;
 pub use medium::*;
 pub use tokenized::*;
-pub use crate::db::orm::*;
+pub use crate::db::models::*;
